clush_rag_project/requirements.txt
langchain
sentence-transformers
faiss-cpu
transformers
torch
peft
datasets
gradio
chromadb
